Clear metadata when unlinking or extending into chunks.
Centralize the clearing and clear the amount required by the size of the
type of chunk in question.

When extending into the designated victim, always clear even though we
may not have internal pointers. This has some cost, but is easier than
trying to figure out if problematic metadata is present.